ref(processor): Move sessions related code into separate sub-module
